+++ libraries/rush-lib/src/logic/operations/CacheableOperationPlugin.ts
@@ -170,14 +170,14 @@ export class CacheableOperationPlugin implements IPhasedCommandPlugin {
                 return operation.name;
               });
 
-              // Generates cluster id, cluster id comes from the project folder and phase name of all operations in the same cluster.
+              // Generates cluster id, cluster id comes from the project folder and operation name of all operations in the same cluster.
               const hash: crypto.Hash = crypto.createHash('sha1');
               for (const operation of groupedOperations) {
                 const { associatedPhase: phase, associatedProject: project } = operation;
                 if (project && phase) {
                   hash.update(project.projectRelativeFolder);
                   hash.update(RushConstants.hashDelimiter);
-                  hash.update(phase.name);
+                  hash.update(operation.name ?? phase.name);
                   hash.update(RushConstants.hashDelimiter);
                 }
               }