+++ libraries/rush-lib/src/logic/test/Git.test.ts
@@ -223,4 +223,41 @@ describe(Git.name, () => {
       await expect(getMockedGitIsRefACommitAsync('HEAD')).resolves.toBe(false);
     });
   });
+
+  describe(Git.prototype.tryGetGitEmailAsync.name, () => {
+    async function getMockGitEmail(hasGitPath: boolean, output: string | Error): Promise<string | undefined> {
+      const gitInstance: Git = new Git({ rushJsonFolder: '/repo/root' } as RushConfiguration);
+      jest.spyOn(gitInstance, 'gitPath', 'get').mockImplementation(() => {
+        if (hasGitPath) return '/git/bin/path';
+        else return undefined;
+      });
+
+      jest
+        .spyOn(gitInstance, '_executeGitCommandAndCaptureOutputAsync')
+        .mockImplementation(async (gitPath: string, args: string[]) => {
+          expect(gitPath).toEqual('/git/bin/path');
+          expect(args).toEqual(['config', 'user.email']);
+          if (typeof output === 'string') return output;
+          else throw output;
+        });
+
+      return await gitInstance.tryGetGitEmailAsync();
+    }
+
+    it('Throw exception when cannot find git path', async () => {
+      await expect(getMockGitEmail(false, 'user@example.com')).rejects.toBeInstanceOf(Error);
+    });
+
+    it('Returns result when git user.email has been found', async () => {
+      await expect(getMockGitEmail(true, 'user@example.com')).resolves.toEqual('user@example.com');
+    });
+
+    it('Returns empty email when git user.email return empty string', async () => {
+      await expect(getMockGitEmail(true, '')).resolves.toEqual('');
+    });
+
+    it('Returns undefined when git user.email not configure', async () => {
+      await expect(getMockGitEmail(true, new Error('Email is missing'))).resolves.toEqual(undefined);
+    });
+  });
 });